requirements.txt
pyyaml
faster-whisper==1.0.0
ctranslate2>=4.3.1
openai-whisper